pyproject.toml
[tool.poetry]
name = "flaskgpt"
version = "0.1.0"
description = "chatGPT Clone using Flask"
authors = ["Yamdev"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0.3"
openai = "^1.40.6"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
